+++ src/biotite/database/entrez/__init__.py
@@ -8,5 +8,5 @@ A subpackage for downloading files from the NCBI Entrez database.
 
 __author__ = "Patrick Kunzmann"
 
-from .fetch import *
-from .search import *
+from .download import *
+from .query import *
